// packages/pwa/_src/Web.types.ts
import { ResizeMode } from '@expo/image-utils';

export type Direction = 'ltr' | 'rtl' | 'auto';

export type Display = 'fullscreen' | 'standalone' | 'minimal-ui' | 'browser';

export type Orientation =
  | 'any'
  | 'natural'
  | 'landscape'
  | 'landscape-primary'
  | 'landscape-secondary'
  | 'portrait'
  | 'portrait-primary'
  | 'portrait-secondary'
  | 'omit';

export type CrossOrigin = 'use-credentials' | 'anonymous';

export interface RelatedApplications {
  platform?: string;
  url: string;
  id?: string;
}

export type Manifest = Partial<{
  background_color: string;
  description: string;
  dir: Direction;
  display: Display;
  lang: string;
  name: string;
  orientation: Orientation;
  prefer_related_applications: boolean;
  related_applications: RelatedApplications[];
  scope: string;
  short_name: string;
  start_url: string;
  theme_color: string;
  crossorigin: CrossOrigin;
}>;

export type Icon = {
  src: string;
  name?: string;
  resizeMode: ResizeMode;
  backgroundColor: string;
  width: number;
  height: number;
  padding?: number;
};

export type WebpackAsset = {
  source: Buffer;
  path: string;
};

export type HtmlTag = {
  tagName: 'link';
  attributes: { rel?: string; href?: string; media?: string; sizes?: string; type?: string };
};

export type SplashIcon = Icon & {
  media: string;
};

export type ProjectOptions = {
  projectRoot: string;
  publicPath: string;
  // unimp
  destination?: string;
};

export type HTMLOutput = { asset: WebpackAsset; tag?: HtmlTag; manifest?: ManifestIcon };

export type IconOptions = Omit<Icon, 'name' | 'width' | 'height'>;
export type ManifestIcon = { src: string; sizes: string; type: 'image/png'; purpose?: string };